PIVOT: Use Cloudflare Workers instead of EC2+nginx. :cloud: :fire:
Add Cloudflare Worker to handle requests in place of nginx on an EC2
instance, removing AWS as a dependency. This requires some custom JS
code, but it gets rid of lots of infrastructure, and it's something to
do. Hook into Datadog for logging just because Cloudflare only gives
access logs to "Enterprise" customers. :pleading_face:

Remove Ansible because there's no server to configure.

Saves :moneybag: too.